by-lapply {doBy}	R Documentation

Formula based version of lapply.

Description

This function is a wrapper for calling lapply on the list resulting from first calling splitBy.

Usage

lapplyBy(formula, data = parent.frame(), FUN)

Arguments

formula
A formula describing how data should be split.

data
A dataframe.

FUN
A function to be applied to each element in the splitted list, see 'Examples' below.

Value

A list.

Author(s)

Søren Højsgaard, sorenh@math.aau.dk

See Also

lapplyBy

Examples

data(dietox)

## Calculate weekwise feed efficiency = weight gain / feed intake
dietox <- orderBy(~Pig + Time, data=dietox)
v <- lapplyBy(~Pig, data=dietox,
              function(d) c(NA, diff(d$Weight) / diff(d$Feed)))
dietox$FE <- unlist(v)

## Technically this is the same as 
dietox <- orderBy(~Pig + Time, data=dietox)
wdata <- splitBy(~Pig, data=dietox)
v <- lapply(wdata, function(d) c(NA, diff(d$Weight)/diff(d$Feed)))
dietox$FE <- unlist(v)

[Package doBy version 4.6-2 Index]