data-personality {doBy}	R Documentation

Personality traits

Description

The 'peronality' dataframe has 240 rows and 32 columns

Usage

data(personality)

Format

This dataframe has recordings on the following 32 variables: distant, talkatv, carelss, hardwrk, anxious, agreebl, tense, kind, opposng, relaxed, disorgn, outgoin, approvn, shy, discipl, harsh, persevr, friendl, worryin, respnsi, contrar, sociabl, lazy, coopera, quiet, organiz, criticl, lax, laidbck, withdrw, givinup, easygon

Author(s)

Søren Højsgaard, sorenh@math.aau.dk

References

Origin unclear

Examples

data(personality)
str(personality)

[Package doBy version 4.6.14 Index]