is-estimable {doBy}	R Documentation

Determines if contrasts are estimable.

Description

Determines if contrasts are estimable, that is, if the contrasts can be written as a linear function of the data.

Usage

is_estimable(K, null.basis)

Arguments

K
A matrix.

null.basis
A basis for a null space (can be found with null_basis()).

Details

Consider the setting E(Y)=Xb. A linear function of b, say l'b is estimable if and only if there exists an r such that r'X=l' or equivalently l=X'r. Hence l must be in the column space of X', i.e. in the orthogonal complement of the null space of X. Hence, with a basis B for the null space, is_estimable() checks if each row l of the matrix K is perpendicular to each column basis vector in B.

Value

A logical vector.

Author(s)

Søren Højsgaard, sorenh@math.aau.dk

References

http://web.mit.edu/18.06/www/Essays/newpaper_ver3.pdf

See Also

null_basis

Examples

## TO BE WRITTEN

[Package doBy version 4.6-2 Index]