summary.esticon_class {doBy}	R Documentation

Compute linear estimates

Description

Compute linear estimates for a range of models. One example of linear estimates is population means (also known as LSMEANS).

Usage

## S3 method for class 'esticon_class'
summary(object, ...)

linest(object, L = NULL, level = 0.95, ...)

## S3 method for class 'linest_class'
coef(object, ...)

## S3 method for class 'linest_class'
summary(object, ...)

## S3 method for class 'linest_class'
confint(object, parm, level = 0.95, ...)

Arguments

object
Model object

...
Additional arguments; currently not used.

L
Either NULL or a matrix with p columns where p is the number of parameters in the systematic effects in the model. If NULL then L is taken to be the p times p identity matrix

level
The level of the (asymptotic) confidence interval.

parm
Specification of the parameters estimates for which confidence inctervals are to be calculated.

Value

A dataframe with results from computing the contrasts.

Author(s)

Søren Højsgaard, sorenh@math.aau.dk

See Also

LSmeans, LE_matrix

Examples

## Make balanced dataset
dat.bal <- expand.grid(list(AA=factor(1:2), BB=factor(1:3), CC=factor(1:3)))
dat.bal$y <- rnorm(nrow(dat.bal))

## Make unbalanced dataset
#   'BB' is nested within 'CC' so BB=1 is only found when CC=1
#   and BB=2,3 are found in each CC=2,3,4
dat.nst <- dat.bal
dat.nst$CC <-factor(c(1,1,2,2,2,2,1,1,3,3,3,3,1,1,4,4,4,4))

mod.bal  <- lm(y ~ AA + BB * CC, data=dat.bal)
mod.nst  <- lm(y ~ AA + BB : CC, data=dat.nst)

L <- LE_matrix(mod.nst, effect=c("BB", "CC"))
linest( mod.nst, L )

[Package doBy version 4.6-2 Index]