mb_summary {doBy}	R Documentation

Fast summary of microbenchmark object

Description

Fast summary of microbenchmark object. The default summary method from the microbenchmark package is fairly slow in producing a summary (due to a call to a function from the multcomp package.)

Usage

mb_summary(object, unit, add.unit = TRUE, ...)

summary_mb(object, unit, add.unit = TRUE, ...)

Arguments

object
A microbenchmark object

unit
The time unit to be used

add.unit
Should time unit be added as column to resulting dataframe.

...
Additional arguments; currently not used.

[Package doBy version 4.6.14 Index]