taylor {doBy}	R Documentation

Taylor expansion (one dimension)

Description

Returns Taylor polynomial approximating a function fn(x)

Usage

taylor(fn, x0, ord = 1)

Arguments

fn
A function of one variable and that variable must be named 'x'.

x0
The point in which to to the Taylor expansion.

ord
The order of the Taylor expansion.

Value

function.

Author(s)

Søren Højsgaard, sorenh@math.aau.dk

Examples

fn <- function(x) log(x)
ord <- 2
x0 <- 2

xv  <- seq(.2, 5, .1)
plot(xv, fn(xv), type="l")
lines(xv, taylor(fn, x0=x0, ord=ord)(xv), lty=2)
abline(v=x0)

fn <- function(x)sin(x)
ord <- 4
x0 <- 0
xv <- seq(-2*pi, 2*pi, 0.1)
plot(xv, fn(xv), type="l")
lines(xv, taylor(fn, x0=x0, ord=ord)(xv), lty=2)
abline(v=x0)

[Package doBy version 4.6.8 Index]