Sampling from a data frame

A data frame is split according to some variables in a formula, and a sample of a certain fraction of each is drawn.

sample_by(data, formula, frac = 0.1, replace = FALSE, systematic = FALSE)

sampleBy(
  formula,
  frac = 0.1,
  replace = FALSE,
  data = parent.frame(),
  systematic = FALSE
)

Arguments

data

A data frame.

formula

A formula defining the grouping of the data frame.

frac

The part of data to be sampled.

replace

Is the sampling with replacement.

systematic

Should sampling be systematic.

Value

A dataframe.

Details

If systematic=FALSE (default) then frac gives the fraction of data sampled. If systematic=TRUE and frac=.2 then every 1/.2 i.e. every 5th observation is taken out.

See also

Examples

data(dietox)
sampleBy(formula = ~ Evit + Cu, frac=.1, data = dietox)
#>                    Pig Litter    Start    Weight       Feed Time
#> Evit000|Cu000.5   4601      1 26.50000  49.09998  45.200001    5
#> Evit000|Cu000.106 4757      3 26.89999  81.00000 129.000000   10
#> Evit000|Cu000.199 4856      6 32.00000  70.00000  75.200001    7
#> Evit000|Cu000.200 4856      6 32.00000  79.09998  91.700005    8
#> Evit000|Cu000.259 5497      8 21.20000  59.29999  68.000000    7
#> Evit000|Cu000.394 5852     12 23.79999  23.79999         NA    1
#> Evit000|Cu000.395 5852     12 23.79999  30.00000   9.700000    2
#> Evit000|Cu000.548 6287     18 21.50000  93.39996 146.599991   11
#> Evit000|Cu035.16  4602      1 28.29999  44.50000  26.100000    4
#> Evit000|Cu035.19  4602      1 28.29999  65.00000  72.800003    7
#> Evit000|Cu035.272 5500      8 17.00000  54.00000  87.000000    8
#> Evit000|Cu035.274 5500      8 17.00000  70.00000 124.800003   10
#> Evit000|Cu035.275 5500      8 17.00000  78.29999 147.300001   11
#> Evit000|Cu035.412 5862     11 22.00000  55.19998  62.000000    7
#> Evit000|Cu035.636 8049     20 22.00000  33.09998  19.900002    3
#> Evit000|Cu035.643 8049     20 22.00000  80.19995 137.600000   10
#> Evit000|Cu035.815 8273     22 25.00000  30.09999   8.000000    2
#> Evit000|Cu035.818 8273     22 25.00000  53.59998  48.500000    5
#> Evit000|Cu175.129 4760      3 15.00000  53.69998  83.800003    9
#> Evit000|Cu175.130 4760      3 15.00000  59.29999  93.200001   10
#> Evit000|Cu175.173 4817      4 24.59999  49.69998  47.000004    5
#> Evit000|Cu175.178 4817      4 24.59999  91.00000 144.000004   10
#> Evit000|Cu175.212 4857      6 26.00000  73.39996 101.800003    8
#> Evit000|Cu175.214 4857      6 26.00000  86.00000 135.300007   10
#> Evit000|Cu175.685 8139     23 24.00000  39.09998  30.300003    4
#> Evit000|Cu175.686 8139     23 24.00000  46.39999  45.100002    5
#> Evit000|Cu175.689 8139     23 24.00000  68.79999  91.800003    8
#> Evit000|Cu175.751 8192     21 35.39999  97.79999 165.000000   10
#> Evit100|Cu000.181 4854      6 29.00000  29.00000         NA    1
#> Evit100|Cu000.188 4854      6 29.00000  70.19995  86.200001    8
#> Evit100|Cu000.310 5524      9 22.39999  78.09998 136.099997   10
#> Evit100|Cu000.479 6058     16 27.39999  32.19998   8.100000    2
#> Evit100|Cu000.488 6058     16 27.39999 100.50000 148.600000   11
#> Evit100|Cu000.515 6211     17 24.50000  31.59999   8.100000    2
#> Evit100|Cu000.524 6211     17 24.50000  96.89996 146.500000   11
#> Evit100|Cu000.529 6284     18 23.09999  40.69998  30.600006    4
#> Evit100|Cu000.591 6433      5 32.19998  60.09998  58.600000    6
#> Evit100|Cu000.596 6433      5 32.19998 111.09998 163.000000   11
#> Evit100|Cu035.38  4605      1 31.50000  34.79999   6.500000    2
#> Evit100|Cu035.220 4858      6 28.00000  44.69998  31.800003    4
#> Evit100|Cu035.225 4858      6 28.00000  83.79999 133.200005    9
#> Evit100|Cu035.227 4858      6 28.00000  97.39996 190.800003   11
#> Evit100|Cu035.585 6432      5 25.09999 106.79999 197.400002   12
#> Evit100|Cu035.732 8191     21 26.59999  38.19998  20.000000    3
#> Evit100|Cu035.740 8191     21 26.59999  97.69995 162.800003   11
#> Evit100|Cu035.828 8437     24 27.00000  35.39999  16.400002    3
#> Evit100|Cu035.831 8437     24 27.00000  56.39999  54.900002    6
#> Evit100|Cu035.833 8437     24 27.00000  75.50000  88.500000    8
#> Evit100|Cu175.50  4641      2 27.09999  33.00000   5.200001    2
#> Evit100|Cu175.51  4641      2 27.09999  42.50000  22.500000    3
#> Evit100|Cu175.54  4641      2 27.09999  63.00000  69.800000    6
#> Evit100|Cu175.163 4815      4 24.89999  65.00000  80.100006    7
#> Evit100|Cu175.166 4815      4 24.89999  83.29999 149.000000   10
#> Evit100|Cu175.313 5527      9 26.59999  28.20000  10.000000    2
#> Evit100|Cu175.315 5527      9 26.59999  45.79999  50.000000    4
#> Evit100|Cu175.434 5866     11 22.00000  42.29999  36.800001    5
#> Evit100|Cu175.435 5866     11 22.00000  50.89999  49.800003    6
#> Evit100|Cu175.438 5866     11 22.00000  71.79999  93.800003    9
#> Evit200|Cu000.96  4756      3 17.00000  89.29999 163.700001   12
#> Evit200|Cu000.324 5528      9 24.00000  28.20000  10.000000    2
#> Evit200|Cu000.326 5528      9 24.00000  42.69998  32.100002    4
#> Evit200|Cu000.351 5581     10 27.00000  57.00000  59.900002    6
#> Evit200|Cu000.352 5581     10 27.00000  63.69998  76.100000    7
#> Evit200|Cu000.376 5850     12 22.79999  53.69998  60.200005    7
#> Evit200|Cu000.490 6207     17 27.09999  27.09999         NA    1
#> Evit200|Cu000.495 6207     17 27.09999  59.00000  57.200001    6
#> Evit200|Cu000.496 6207     17 27.09999  67.29999  72.100002    7
#> Evit200|Cu000.501 6207     17 27.09999 101.39996 153.800003   12
#> Evit200|Cu035.77  4645      2 27.70000  54.79999  55.200001    5
#> Evit200|Cu035.109 4759      3 22.59999  22.59999         NA    1
#> Evit200|Cu035.110 4759      3 22.59999  28.50000  10.600006    2
#> Evit200|Cu035.113 4759      3 22.59999  47.59998  47.100000    5
#> Evit200|Cu035.249 5392      7 24.59999  63.00000  89.499996    9
#> Evit200|Cu035.601 6909     13 24.70000  42.19998  26.600000    4
#> Evit200|Cu035.602 6909     13 24.70000  48.19998  38.000000    5
#> Evit200|Cu035.607 6909     13 24.70000  87.00000 116.900002   10
#> Evit200|Cu035.754 8193     21 29.29999  29.29999         NA    1
#> Evit200|Cu035.758 8193     21 29.29999  51.79999  51.000000    5
#> Evit200|Cu035.760 8193     21 29.29999  62.59998  86.200001    7
#> Evit200|Cu175.31  4603      1 27.59999  68.50000  80.500000    7
#> Evit200|Cu175.35  4603      1 27.59999  98.39996 145.000000   11
#> Evit200|Cu175.239 5389      7 24.29999  87.50000 131.100004   11
#> Evit200|Cu175.365 5582     10 29.50000  72.39996  86.800001    8
#> Evit200|Cu175.366 5582     10 29.50000  80.19995 105.200001    9
#> Evit200|Cu175.367 5582     10 29.50000  90.39996 127.800001   10
#> Evit200|Cu175.369 5582     10 29.50000  99.09998 167.500000   12
#> Evit200|Cu175.460 6056     16 32.50000  76.59998  89.000000    7
#> Evit200|Cu175.678 8053     20 21.89999  80.59998 138.000000    9
#> Evit200|Cu175.841 8439     24 27.29999  43.89999  33.200001    4