Finds subsets of a dataframe which is split by variables in a formula.

A data frame is split by a formula into groups. Then subsets are found within each group, and the result is collected into a data frame.

subset_by(data, formula, subset, select, drop = FALSE, join = TRUE, ...)

subsetBy(
  formula,
  subset,
  data = parent.frame(),
  select,
  drop = FALSE,
  join = TRUE,
  ...
)

Arguments

data

A data frame.

formula

A right hand sided formula or a character vector of variables to split by.

subset

logical expression indicating elements or rows to keep: missing values are taken as false.

select

expression, indicating columns to select from a data frame.

drop

passed on to [ indexing operator.

join

If FALSE the result is a list of data frames (as defined by 'formula'); if TRUE one data frame is returned.

...

further arguments to be passed to or from other methods.

Value

A data frame.

See also

Author

Søren Højsgaard, sorenh@math.aau.dk

Examples

data(dietox)
subsetBy(~Evit, Weight < mean(Weight), data=dietox)
#>              Pig    Cu Litter    Start   Weight      Feed Time
#> Evit000.1   4601 Cu000      1 26.50000 26.50000        NA    1
#> Evit000.2   4601 Cu000      1 26.50000 27.59999  5.200005    2
#> Evit000.3   4601 Cu000      1 26.50000 36.50000 17.600000    3
#> Evit000.4   4601 Cu000      1 26.50000 40.29999 28.500000    4
#> Evit000.5   4601 Cu000      1 26.50000 49.09998 45.200001    5
#> Evit000.6   4601 Cu000      1 26.50000 55.39999 56.900002    6
#> Evit000.7   4601 Cu000      1 26.50000 59.59998 71.700005    7
#> Evit000.13  4602 Cu035      1 28.29999 28.29999        NA    1
#> Evit000.14  4602 Cu035      1 28.29999 30.09999  3.300003    2
#> Evit000.15  4602 Cu035      1 28.29999 38.29999 13.200001    3
#> Evit000.16  4602 Cu035      1 28.29999 44.50000 26.100000    4
#> Evit000.17  4602 Cu035      1 28.29999 51.59998 43.600002    5
#> Evit000.18  4602 Cu035      1 28.29999 57.59998 55.200001    6
#> Evit000.61  4643 Cu000      2 27.00000 27.00000        NA    1
#> Evit000.62  4643 Cu000      2 27.00000 31.79999  6.400002    2
#> Evit000.63  4643 Cu000      2 27.00000 39.00000 21.500000    3
#> Evit000.64  4643 Cu000      2 27.00000 44.79999 33.900002    4
#> Evit000.65  4643 Cu000      2 27.00000 50.89999 44.900002    5
#> Evit000.66  4643 Cu000      2 27.00000 57.39999 58.800000    6
#> Evit000.97  4757 Cu000      3 26.89999 26.89999        NA    1
#> Evit000.98  4757 Cu000      3 26.89999 32.29999  9.900002    2
#> Evit000.99  4757 Cu000      3 26.89999 38.59998 20.200005    3
#> Evit000.100 4757 Cu000      3 26.89999 46.00000 30.000000    4
#> Evit000.101 4757 Cu000      3 26.89999 57.79999 40.000000    5
#> Evit000.102 4757 Cu000      3 26.89999 58.59998 62.800003    6
#> Evit000.121 4760 Cu175      3 15.00000 15.00000        NA    1
#> Evit000.122 4760 Cu175      3 15.00000 19.39999  7.900002    2
#> Evit000.123 4760 Cu175      3 15.00000 21.39999 15.000000    3
#> Evit000.124 4760 Cu175      3 15.00000 25.29999 22.100002    4
#> Evit000.125 4760 Cu175      3 15.00000 31.89999 29.900002    5
#> Evit000.126 4760 Cu175      3 15.00000 37.59998 41.800003    6
#> Evit000.127 4760 Cu175      3 15.00000 43.69998 57.600006    7
#> Evit000.128 4760 Cu175      3 15.00000 48.89999 69.500000    8
#> Evit000.129 4760 Cu175      3 15.00000 53.69998 83.800003    9
#> Evit000.130 4760 Cu175      3 15.00000 59.29999 93.200001   10
#> Evit000.133 4813 Cu035      4 27.29999 27.29999        NA    1
#> Evit000.134 4813 Cu035      4 27.29999 32.39999 10.300003    2
#> Evit000.135 4813 Cu035      4 27.29999 38.29999 22.800003    3
#> Evit000.136 4813 Cu035      4 27.29999 45.79999 40.200005    4
#> Evit000.137 4813 Cu035      4 27.29999 53.79999 53.700005    5
#> Evit000.138 4813 Cu035      4 27.29999 60.00000 72.700005    6
#> Evit000.169 4817 Cu175      4 24.59999 24.59999        NA    1
#> Evit000.170 4817 Cu175      4 24.59999 28.00000  9.000000    2
#> Evit000.171 4817 Cu175      4 24.59999 34.59998 19.900002    3
#> Evit000.172 4817 Cu175      4 24.59999 41.50000 33.900002    4
#> Evit000.173 4817 Cu175      4 24.59999 49.69998 47.000004    5
#> Evit000.174 4817 Cu175      4 24.59999 56.50000 65.200001    6
#> Evit000.193 4856 Cu000      6 32.00000 32.00000        NA    1
#> Evit000.194 4856 Cu000      6 32.00000 35.19998  8.100000    2
#> Evit000.195 4856 Cu000      6 32.00000 41.50000 19.600006    3
#> Evit000.196 4856 Cu000      6 32.00000 47.69998 27.600000    4
#> Evit000.197 4856 Cu000      6 32.00000 53.79999 41.800003    5
#> Evit000.205 4857 Cu175      6 26.00000 26.00000        NA    1
#> Evit000.206 4857 Cu175      6 26.00000 34.50000  8.500000    2
#> Evit000.207 4857 Cu175      6 26.00000 40.39999 19.200005    3
#> Evit000.208 4857 Cu175      6 26.00000 45.00000 29.000000    4
#> Evit000.209 4857 Cu175      6 26.00000 53.50000 45.000000    5
#> Evit000.210 4857 Cu175      6 26.00000 60.19998 62.300003    6
#> Evit000.253 5497 Cu000      8 21.20000 21.20000        NA    1
#> Evit000.254 5497 Cu000      8 21.20000 25.79999  6.300001    2
#> Evit000.255 5497 Cu000      8 21.20000 29.39999 16.500000    3
#> Evit000.256 5497 Cu000      8 21.20000 36.79999 27.200001    4
#> Evit000.257 5497 Cu000      8 21.20000 43.19998 40.200005    5
#> Evit000.258 5497 Cu000      8 21.20000 51.09998 54.800003    6
#> Evit000.259 5497 Cu000      8 21.20000 59.29999 68.000000    7
#> Evit000.265 5500 Cu035      8 17.00000 17.00000        NA    1
#> Evit000.266 5500 Cu035      8 17.00000 19.00000  5.200001    2
#> Evit000.267 5500 Cu035      8 17.00000 23.00000 18.500000    3
#> Evit000.268 5500 Cu035      8 17.00000 29.39999 32.800003    4
#> Evit000.269 5500 Cu035      8 17.00000 34.59998 43.800003    5
#> Evit000.270 5500 Cu035      8 17.00000 40.59998 55.800003    6
#> Evit000.271 5500 Cu035      8 17.00000 46.59998 71.400002    7
#> Evit000.272 5500 Cu035      8 17.00000 54.00000 87.000000    8
#> Evit000.394 5852 Cu000     12 23.79999 23.79999        NA    1
#> Evit000.395 5852 Cu000     12 23.79999 30.00000  9.700000    2
#> Evit000.396 5852 Cu000     12 23.79999 37.09998 20.800003    3
#> Evit000.397 5852 Cu000     12 23.79999 43.19998 31.200005    4
#> Evit000.398 5852 Cu000     12 23.79999 50.39999 44.900002    5
#> Evit000.399 5852 Cu000     12 23.79999 60.19998 61.800003    6
#> Evit000.406 5862 Cu035     11 22.00000 22.00000        NA    1
#> Evit000.407 5862 Cu035     11 22.00000 25.79999  8.500000    2
#> Evit000.408 5862 Cu035     11 22.00000 29.20000 15.800003    3
#> Evit000.409 5862 Cu035     11 22.00000 35.79999 25.900002    4
#> Evit000.410 5862 Cu035     11 22.00000 41.09998 37.600000    5
#> Evit000.411 5862 Cu035     11 22.00000 50.00000 48.200005    6
#> Evit000.412 5862 Cu035     11 22.00000 55.19998 62.000000    7
#> Evit000.502 6208 Cu035     17 26.79999 26.79999        NA    1
#> Evit000.503 6208 Cu035     17 26.79999 31.09999  6.100000    2
#> Evit000.504 6208 Cu035     17 26.79999 37.79999 14.900002    3
#> Evit000.505 6208 Cu035     17 26.79999 44.59998 28.500000    4
#> Evit000.506 6208 Cu035     17 26.79999 50.50000 40.500000    5
#> Evit000.507 6208 Cu035     17 26.79999 58.29999 55.200001    6
#> Evit000.538 6287 Cu000     18 21.50000 21.50000        NA    1
#> Evit000.539 6287 Cu000     18 21.50000 24.59999  5.000000    2
#> Evit000.540 6287 Cu000     18 21.50000 28.50000 14.200001    3
#> Evit000.541 6287 Cu000     18 21.50000 39.09998 27.800003    4
#> Evit000.542 6287 Cu000     18 21.50000 45.29999 39.800003    5
#> Evit000.543 6287 Cu000     18 21.50000 53.00000 53.500000    6
#> Evit000.544 6287 Cu000     18 21.50000 59.59998 68.500000    7
#> Evit000.634 8049 Cu035     20 22.00000 22.00000        NA    1
#> Evit000.635 8049 Cu035     20 22.00000 26.59999  9.500000    2
#> Evit000.636 8049 Cu035     20 22.00000 33.09998 19.900002    3
#> Evit000.637 8049 Cu035     20 22.00000 38.39999 31.000000    4
#> Evit000.638 8049 Cu035     20 22.00000 46.59998 44.900002    5
#> Evit000.639 8049 Cu035     20 22.00000 52.29999 63.800003    6
#> Evit000.646 8050 Cu175     20 26.79999 26.79999        NA    1
#> Evit000.647 8050 Cu175     20 26.79999 33.59998 10.900002    2
#> Evit000.648 8050 Cu175     20 26.79999 42.09998 25.100002    3
#> Evit000.649 8050 Cu175     20 26.79999 47.89999 38.000000    4
#> Evit000.650 8050 Cu175     20 26.79999 56.19998 54.400002    5
#> Evit000.682 8139 Cu175     23 24.00000 24.00000        NA    1
#> Evit000.683 8139 Cu175     23 24.00000 26.79999  6.000000    2
#> Evit000.684 8139 Cu175     23 24.00000 33.00000 20.000000    3
#> Evit000.685 8139 Cu175     23 24.00000 39.09998 30.300003    4
#> Evit000.686 8139 Cu175     23 24.00000 46.39999 45.100002    5
#> Evit000.687 8139 Cu175     23 24.00000 54.00000 61.400002    6
#> Evit000.706 8142 Cu035     23 25.20000 25.20000        NA    1
#> Evit000.707 8142 Cu035     23 25.20000 29.79999  9.000000    2
#> Evit000.708 8142 Cu035     23 25.20000 35.69998 21.000000    3
#> Evit000.709 8142 Cu035     23 25.20000 42.00000 31.300003    4
#> Evit000.710 8142 Cu035     23 25.20000 50.00000 44.000000    5
#> Evit000.711 8142 Cu035     23 25.20000 55.89999 57.800001    6
#> Evit000.742 8192 Cu175     21 35.39999 35.39999        NA    1
#> Evit000.743 8192 Cu175     21 35.39999 42.29999 15.100002    2
#> Evit000.744 8192 Cu175     21 35.39999 49.50000 29.800003    3
#> Evit000.745 8192 Cu175     21 35.39999 58.69998 48.500000    4
#> Evit000.790 8270 Cu175     22 23.70000 23.70000        NA    1
#> Evit000.791 8270 Cu175     22 23.70000 28.79999  7.600000    2
#> Evit000.792 8270 Cu175     22 23.70000 35.09998 18.500000    3
#> Evit000.793 8270 Cu175     22 23.70000 43.00000 33.200001    4
#> Evit000.794 8270 Cu175     22 23.70000 49.09998 44.900002    5
#> Evit000.795 8270 Cu175     22 23.70000 58.79999 60.000000    6
#> Evit000.814 8273 Cu035     22 25.00000 25.00000        NA    1
#> Evit000.815 8273 Cu035     22 25.00000 30.09999  8.000000    2
#> Evit000.816 8273 Cu035     22 25.00000 37.09998 20.000000    3
#> Evit000.817 8273 Cu035     22 25.00000 46.00000 33.800003    4
#> Evit000.818 8273 Cu035     22 25.00000 53.59998 48.500000    5
#> Evit000.850 8442 Cu175     24 25.70000 25.70000        NA    1
#> Evit000.851 8442 Cu175     24 25.70000 28.70000  6.600000    2
#> Evit000.852 8442 Cu175     24 25.70000 33.39999 15.800003    3
#> Evit000.853 8442 Cu175     24 25.70000 40.00000 27.000000    4
#> Evit000.854 8442 Cu175     24 25.70000 46.69998 40.000000    5
#> Evit000.855 8442 Cu175     24 25.70000 56.59998 52.000000    6
#> Evit100.37  4605 Cu035      1 31.50000 31.50000        NA    1
#> Evit100.38  4605 Cu035      1 31.50000 34.79999  6.500000    2
#> Evit100.39  4605 Cu035      1 31.50000 40.69998 20.000000    3
#> Evit100.40  4605 Cu035      1 31.50000 47.69998 33.100002    4
#> Evit100.41  4605 Cu035      1 31.50000 55.89999 51.800003    5
#> Evit100.42  4605 Cu035      1 31.50000 62.19998 65.800003    6
#> Evit100.49  4641 Cu175      2 27.09999 27.09999        NA    1
#> Evit100.50  4641 Cu175      2 27.09999 33.00000  5.200001    2
#> Evit100.51  4641 Cu175      2 27.09999 42.50000 22.500000    3
#> Evit100.52  4641 Cu175      2 27.09999 50.09998 38.500000    4
#> Evit100.53  4641 Cu175      2 27.09999 56.50000 53.300003    5
#> Evit100.145 4814 Cu035      4 26.20000 26.20000        NA    1
#> Evit100.146 4814 Cu035      4 26.20000 31.09999  9.900002    2
#> Evit100.147 4814 Cu035      4 26.20000 34.59998 19.600006    3
#> Evit100.148 4814 Cu035      4 26.20000 41.19998 30.000000    4
#> Evit100.149 4814 Cu035      4 26.20000 49.19998 37.400002    5
#> Evit100.150 4814 Cu035      4 26.20000 55.59998 57.200005    6
#> Evit100.157 4815 Cu175      4 24.89999 24.89999        NA    1
#> Evit100.158 4815 Cu175      4 24.89999 29.70000  9.500000    2
#> Evit100.159 4815 Cu175      4 24.89999 33.79999 19.000000    3
#> Evit100.160 4815 Cu175      4 24.89999 38.79999 32.000000    4
#> Evit100.161 4815 Cu175      4 24.89999 48.29999 45.500000    5
#> Evit100.162 4815 Cu175      4 24.89999 52.89999 59.100006    6
#> Evit100.181 4854 Cu000      6 29.00000 29.00000        NA    1
#> Evit100.182 4854 Cu000      6 29.00000 32.69998  8.200001    2
#> Evit100.183 4854 Cu000      6 29.00000 38.29999 19.800003    3
#> Evit100.184 4854 Cu000      6 29.00000 45.50000 30.000000    4
#> Evit100.185 4854 Cu000      6 29.00000 48.59998 40.000000    5
#> Evit100.186 4854 Cu000      6 29.00000 56.00000 50.000000    6
#> Evit100.217 4858 Cu035      6 28.00000 28.00000        NA    1
#> Evit100.218 4858 Cu035      6 28.00000 31.20000  8.000000    2
#> Evit100.219 4858 Cu035      6 28.00000 38.79999 16.700005    3
#> Evit100.220 4858 Cu035      6 28.00000 44.69998 31.800003    4
#> Evit100.221 4858 Cu035      6 28.00000 53.59998 50.500004    5
#> Evit100.222 4858 Cu035      6 28.00000 59.19998 71.100006    6
#> Evit100.277 5501 Cu175      8 30.00000 30.00000        NA    1
#> Evit100.278 5501 Cu175      8 30.00000 34.69998 10.000000    2
#> Evit100.279 5501 Cu175      8 30.00000 42.79999 22.500000    3
#> Evit100.280 5501 Cu175      8 30.00000 48.39999 35.200001    4
#> Evit100.281 5501 Cu175      8 30.00000 53.29999 45.800003    5
#> Evit100.282 5501 Cu175      8 30.00000 60.00000 60.000000    6
#> Evit100.301 5524 Cu000      9 22.39999 22.39999        NA    1
#> Evit100.302 5524 Cu000      9 22.39999 25.20000 10.000000    2
#> Evit100.303 5524 Cu000      9 22.39999 30.00000 26.699997    3
#> Evit100.304 5524 Cu000      9 22.39999 36.50000 39.799999    4
#> Evit100.305 5524 Cu000      9 22.39999 43.29999 55.199997    5
#> Evit100.306 5524 Cu000      9 22.39999 50.00000 74.499996    6
#> Evit100.307 5524 Cu000      9 22.39999 58.09998 85.999996    7
#> Evit100.312 5527 Cu175      9 26.59999 26.59999        NA    1
#> Evit100.313 5527 Cu175      9 26.59999 28.20000 10.000000    2
#> Evit100.314 5527 Cu175      9 26.59999 38.39999 30.000000    3
#> Evit100.315 5527 Cu175      9 26.59999 45.79999 50.000000    4
#> Evit100.316 5527 Cu175      9 26.59999 54.89999 60.000000    5
#> Evit100.317 5527 Cu175      9 26.59999 60.09998 78.500000    6
#> Evit100.318 5527 Cu175      9 26.59999 56.89999 86.200001    7
#> Evit100.334 5578 Cu175     10 27.09999 27.09999        NA    1
#> Evit100.335 5578 Cu175     10 27.09999 34.69998 11.100002    2
#> Evit100.336 5578 Cu175     10 27.09999 40.69998 29.900002    3
#> Evit100.337 5578 Cu175     10 27.09999 47.59998 45.300001    4
#> Evit100.338 5578 Cu175     10 27.09999 56.50000 65.600000    5
#> Evit100.430 5866 Cu175     11 22.00000 22.00000        NA    1
#> Evit100.431 5866 Cu175     11 22.00000 24.79999  7.500000    2
#> Evit100.432 5866 Cu175     11 22.00000 30.00000 16.100000    3
#> Evit100.433 5866 Cu175     11 22.00000 35.79999 26.000000    4
#> Evit100.434 5866 Cu175     11 22.00000 42.29999 36.800001    5
#> Evit100.435 5866 Cu175     11 22.00000 50.89999 49.800003    6
#> Evit100.436 5866 Cu175     11 22.00000 57.79999 65.100002    7
#> Evit100.466 6057 Cu175     16 27.00000 27.00000        NA    1
#> Evit100.467 6057 Cu175     16 27.00000 32.39999  7.800001    2
#> Evit100.468 6057 Cu175     16 27.00000 38.19998 17.000000    3
#> Evit100.469 6057 Cu175     16 27.00000 44.69998 30.100006    4
#> Evit100.470 6057 Cu175     16 27.00000 53.19998 44.800003    5
#> Evit100.478 6058 Cu000     16 27.39999 27.39999        NA    1
#> Evit100.479 6058 Cu000     16 27.39999 32.19998  8.100000    2
#> Evit100.480 6058 Cu000     16 27.39999 37.89999 18.300000    3
#> Evit100.481 6058 Cu000     16 27.39999 45.89999 29.900002    4
#> Evit100.482 6058 Cu000     16 27.39999 51.19998 44.100002    5
#> Evit100.483 6058 Cu000     16 27.39999 57.09998 58.300000    6
#> Evit100.514 6211 Cu000     17 24.50000 24.50000        NA    1
#> Evit100.515 6211 Cu000     17 24.50000 31.59999  8.100000    2
#> Evit100.516 6211 Cu000     17 24.50000 37.39999 18.100000    3
#> Evit100.517 6211 Cu000     17 24.50000 46.39999 30.100006    4
#> Evit100.518 6211 Cu000     17 24.50000 53.59998 48.100000    5
#> Evit100.519 6211 Cu000     17 24.50000 61.19998 62.600002    6
#> Evit100.526 6284 Cu000     18 23.09999 23.09999        NA    1
#> Evit100.527 6284 Cu000     18 23.09999 26.00000  5.900002    2
#> Evit100.528 6284 Cu000     18 23.09999 30.09999 17.600000    3
#> Evit100.529 6284 Cu000     18 23.09999 40.69998 30.600006    4
#> Evit100.530 6284 Cu000     18 23.09999 46.59998 41.400002    5
#> Evit100.531 6284 Cu000     18 23.09999 52.09998 53.400002    6
#> Evit100.532 6284 Cu000     18 23.09999 60.50000 69.200000    7
#> Evit100.562 6430 Cu175      5 34.19998 34.19998        NA    1
#> Evit100.563 6430 Cu175      5 34.19998 42.09998  8.499996    2
#> Evit100.564 6430 Cu175      5 34.19998 50.09998 23.099998    3
#> Evit100.565 6430 Cu175      5 34.19998 57.59998 35.000000    4
#> Evit100.566 6430 Cu175      5 34.19998 62.29999 54.299995    5
#> Evit100.574 6432 Cu035      5 25.09999 25.09999        NA    1
#> Evit100.575 6432 Cu035      5 25.09999 23.20000  4.000000    2
#> Evit100.576 6432 Cu035      5 25.09999 28.89999 11.100006    3
#> Evit100.577 6432 Cu035      5 25.09999 36.79999 21.000000    4
#> Evit100.578 6432 Cu035      5 25.09999 44.19998 31.000000    5
#> Evit100.579 6432 Cu035      5 25.09999 52.79999 56.300003    6
#> Evit100.580 6432 Cu035      5 25.09999 62.19998 69.100006    7
#> Evit100.586 6433 Cu000      5 32.19998 32.19998        NA    1
#> Evit100.587 6433 Cu000      5 32.19998 36.39999  6.900002    2
#> Evit100.588 6433 Cu000      5 32.19998 41.79999 14.300003    3
#> Evit100.589 6433 Cu000      5 32.19998 48.19998 26.400002    4
#> Evit100.590 6433 Cu000      5 32.19998 55.89999 41.800003    5
#> Evit100.591 6433 Cu000      5 32.19998 60.09998 58.600000    6
#> Evit100.610 6910 Cu000     13 24.20000 24.20000        NA    1
#> Evit100.611 6910 Cu000     13 24.20000 29.20000  6.200001    2
#> Evit100.612 6910 Cu000     13 24.20000 36.19998 17.100000    3
#> Evit100.613 6910 Cu000     13 24.20000 43.09998 30.000000    4
#> Evit100.614 6910 Cu000     13 24.20000 49.19998 44.000000    5
#> Evit100.615 6910 Cu000     13 24.20000 57.00000 56.900002    6
#> Evit100.658 8051 Cu035     20 30.59999 30.59999        NA    1
#> Evit100.659 8051 Cu035     20 30.59999 37.59998 11.200005    2
#> Evit100.660 8051 Cu035     20 30.59999 42.79999 21.000000    3
#> Evit100.661 8051 Cu035     20 30.59999 47.39999 31.100006    4
#> Evit100.662 8051 Cu035     20 30.59999 56.59998 47.000000    5
#> Evit100.718 8144 Cu035     23 23.50000 23.50000        NA    1
#> Evit100.719 8144 Cu035     23 23.50000 25.89999  7.200001    2
#> Evit100.720 8144 Cu035     23 23.50000 32.19998 19.500000    3
#> Evit100.721 8144 Cu035     23 23.50000 32.39999 25.800003    4
#> Evit100.722 8144 Cu035     23 23.50000 29.79999 29.500000    5
#> Evit100.723 8144 Cu035     23 23.50000 32.79999 34.000000    6
#> Evit100.724 8144 Cu035     23 23.50000 41.00000 50.000000    7
#> Evit100.725 8144 Cu035     23 23.50000 50.00000 66.700001    8
#> Evit100.726 8144 Cu035     23 23.50000 56.89999 80.700005    9
#> Evit100.730 8191 Cu035     21 26.59999 26.59999        NA    1
#> Evit100.731 8191 Cu035     21 26.59999 33.00000 10.000000    2
#> Evit100.732 8191 Cu035     21 26.59999 38.19998 20.000000    3
#> Evit100.733 8191 Cu035     21 26.59999 46.39999 30.000000    4
#> Evit100.734 8191 Cu035     21 26.59999 53.19998 40.000000    5
#> Evit100.735 8191 Cu035     21 26.59999 59.59998 60.000000    6
#> Evit100.802 8271 Cu000     22 24.00000 24.00000        NA    1
#> Evit100.803 8271 Cu000     22 24.00000 28.89999  8.500000    2
#> Evit100.804 8271 Cu000     22 24.00000 35.79999 19.500000    3
#> Evit100.805 8271 Cu000     22 24.00000 44.00000 30.000000    4
#> Evit100.806 8271 Cu000     22 24.00000 47.50000 43.200001    5
#> Evit100.807 8271 Cu000     22 24.00000 56.39999 56.600000    6
#> Evit100.826 8437 Cu035     24 27.00000 27.00000        NA    1
#> Evit100.827 8437 Cu035     24 27.00000 30.00000  6.800003    2
#> Evit100.828 8437 Cu035     24 27.00000 35.39999 16.400002    3
#> Evit100.829 8437 Cu035     24 27.00000 41.00000 28.300001    4
#> Evit100.830 8437 Cu035     24 27.00000 48.59998 41.000000    5
#> Evit100.831 8437 Cu035     24 27.00000 56.39999 54.900002    6
#> Evit200.25  4603 Cu175      1 27.59999 27.59999        NA    1
#> Evit200.26  4603 Cu175      1 27.59999 30.59999  6.700005    2
#> Evit200.27  4603 Cu175      1 27.59999 38.69998 17.600000    3
#> Evit200.28  4603 Cu175      1 27.59999 47.19998 32.800003    4
#> Evit200.29  4603 Cu175      1 27.59999 54.09998 50.000000    5
#> Evit200.73  4645 Cu035      2 27.70000 27.70000        NA    1
#> Evit200.74  4645 Cu035      2 27.70000 33.59998 10.000000    2
#> Evit200.75  4645 Cu035      2 27.70000 44.00000 28.500000    3
#> Evit200.76  4645 Cu035      2 27.70000 46.69998 40.000000    4
#> Evit200.77  4645 Cu035      2 27.70000 54.79999 55.200001    5
#> Evit200.85  4756 Cu000      3 17.00000 17.00000        NA    1
#> Evit200.86  4756 Cu000      3 17.00000 19.00000  9.900002    2
#> Evit200.87  4756 Cu000      3 17.00000 23.59999 18.500000    3
#> Evit200.88  4756 Cu000      3 17.00000 30.00000 28.900002    4
#> Evit200.89  4756 Cu000      3 17.00000 36.29999 38.100002    5
#> Evit200.90  4756 Cu000      3 17.00000 42.39999 52.799999    6
#> Evit200.91  4756 Cu000      3 17.00000 50.50000 66.799999    7
#> Evit200.92  4756 Cu000      3 17.00000 57.59998 86.000000    8
#> Evit200.109 4759 Cu035      3 22.59999 22.59999        NA    1
#> Evit200.110 4759 Cu035      3 22.59999 28.50000 10.600006    2
#> Evit200.111 4759 Cu035      3 22.59999 32.69998 21.100006    3
#> Evit200.112 4759 Cu035      3 22.59999 39.89999 33.300003    4
#> Evit200.113 4759 Cu035      3 22.59999 47.59998 47.100000    5
#> Evit200.114 4759 Cu035      3 22.59999 53.79999 63.400002    6
#> Evit200.229 5389 Cu175      7 24.29999 24.29999        NA    1
#> Evit200.230 5389 Cu175      7 24.29999 28.39999  5.400002    2
#> Evit200.231 5389 Cu175      7 24.29999 36.00000 20.000000    3
#> Evit200.232 5389 Cu175      7 24.29999 40.29999 30.300003    4
#> Evit200.233 5389 Cu175      7 24.29999 46.19998 41.800003    5
#> Evit200.234 5389 Cu175      7 24.29999 51.89999 54.800003    6
#> Evit200.235 5389 Cu175      7 24.29999 58.09998 69.000002    7
#> Evit200.241 5392 Cu035      7 24.59999 24.59999        NA    1
#> Evit200.242 5392 Cu035      7 24.59999 25.79999  7.700001    2
#> Evit200.243 5392 Cu035      7 24.59999 28.59999 19.999996    3
#> Evit200.244 5392 Cu035      7 24.59999 30.39999 29.999996    4
#> Evit200.245 5392 Cu035      7 24.59999 36.09998 39.999996    5
#> Evit200.246 5392 Cu035      7 24.59999 36.09998 47.499996    6
#> Evit200.247 5392 Cu035      7 24.59999 49.29999 62.499996    7
#> Evit200.248 5392 Cu035      7 24.59999 57.89999 76.799999    8
#> Evit200.289 5502 Cu000      8 29.59999 29.59999        NA    1
#> Evit200.290 5502 Cu000      8 29.59999 34.19998  9.000000    2
#> Evit200.291 5502 Cu000      8 29.59999 43.00000 21.400002    3
#> Evit200.292 5502 Cu000      8 29.59999 50.00000 36.500000    4
#> Evit200.293 5502 Cu000      8 29.59999 57.00000 51.000000    5
#> Evit200.323 5528 Cu000      9 24.00000 24.00000        NA    1
#> Evit200.324 5528 Cu000      9 24.00000 28.20000 10.000000    2
#> Evit200.325 5528 Cu000      9 24.00000 35.29999 20.000000    3
#> Evit200.326 5528 Cu000      9 24.00000 42.69998 32.100002    4
#> Evit200.327 5528 Cu000      9 24.00000 48.19998 44.800003    5
#> Evit200.328 5528 Cu000      9 24.00000 53.19998 58.500000    6
#> Evit200.346 5581 Cu000     10 27.00000 27.00000        NA    1
#> Evit200.347 5581 Cu000     10 27.00000 30.00000  6.800001    2
#> Evit200.348 5581 Cu000     10 27.00000 37.79999 19.500000    3
#> Evit200.349 5581 Cu000     10 27.00000 43.29999 31.100006    4
#> Evit200.350 5581 Cu000     10 27.00000 51.50000 46.200001    5
#> Evit200.351 5581 Cu000     10 27.00000 57.00000 59.900002    6
#> Evit200.358 5582 Cu175     10 29.50000 29.50000        NA    1
#> Evit200.359 5582 Cu175     10 29.50000 31.39999  6.800001    2
#> Evit200.360 5582 Cu175     10 29.50000 39.39999 19.600000    3
#> Evit200.361 5582 Cu175     10 29.50000 44.79999 30.300003    4
#> Evit200.362 5582 Cu175     10 29.50000 50.29999 41.400002    5
#> Evit200.363 5582 Cu175     10 29.50000 57.00000 55.000000    6
#> Evit200.370 5850 Cu000     12 22.79999 22.79999        NA    1
#> Evit200.371 5850 Cu000     12 22.79999 27.00000  6.000000    2
#> Evit200.372 5850 Cu000     12 22.79999 33.89999 18.200001    3
#> Evit200.373 5850 Cu000     12 22.79999 37.69998 27.000000    4
#> Evit200.374 5850 Cu000     12 22.79999 41.39999 36.300001    5
#> Evit200.375 5850 Cu000     12 22.79999 48.19998 48.000000    6
#> Evit200.376 5850 Cu000     12 22.79999 53.69998 60.200005    7
#> Evit200.382 5851 Cu175     12 22.29999 22.29999        NA    1
#> Evit200.383 5851 Cu175     12 22.29999 24.70000  5.900002    2
#> Evit200.384 5851 Cu175     12 22.29999 30.39999 15.000000    3
#> Evit200.385 5851 Cu175     12 22.29999 36.89999 25.200001    4
#> Evit200.386 5851 Cu175     12 22.29999 44.09998 39.000000    5
#> Evit200.387 5851 Cu175     12 22.29999 52.79999 56.000000    6
#> Evit200.418 5865 Cu035     11 22.00000 22.00000        NA    1
#> Evit200.419 5865 Cu035     11 22.00000 27.00000  8.800000    2
#> Evit200.420 5865 Cu035     11 22.00000 32.89999 18.900001    3
#> Evit200.421 5865 Cu035     11 22.00000 38.00000 29.200005    4
#> Evit200.422 5865 Cu035     11 22.00000 46.19998 43.600002    5
#> Evit200.423 5865 Cu035     11 22.00000 53.50000 58.100000    6
#> Evit200.424 5865 Cu035     11 22.00000 58.59998 73.400002    7
#> Evit200.442 6055 Cu035     16 26.20000 26.20000        NA    1
#> Evit200.443 6055 Cu035     16 26.20000 30.39999  6.900002    2
#> Evit200.444 6055 Cu035     16 26.20000 37.39999 18.200001    3
#> Evit200.445 6055 Cu035     16 26.20000 43.00000 30.600006    4
#> Evit200.446 6055 Cu035     16 26.20000 49.50000 42.100002    5
#> Evit200.447 6055 Cu035     16 26.20000 55.09998 62.200001    6
#> Evit200.454 6056 Cu175     16 32.50000 32.50000        NA    1
#> Evit200.455 6056 Cu175     16 32.50000 38.79999 10.000000    2
#> Evit200.456 6056 Cu175     16 32.50000 47.79999 23.800003    3
#> Evit200.457 6056 Cu175     16 32.50000 54.50000 39.000000    4
#> Evit200.490 6207 Cu000     17 27.09999 27.09999        NA    1
#> Evit200.491 6207 Cu000     17 27.09999 30.29999  5.800003    2
#> Evit200.492 6207 Cu000     17 27.09999 38.19998 17.000000    3
#> Evit200.493 6207 Cu000     17 27.09999 43.19998 30.000000    4
#> Evit200.494 6207 Cu000     17 27.09999 52.50000 45.000000    5
#> Evit200.495 6207 Cu000     17 27.09999 59.00000 57.200001    6
#> Evit200.550 6288 Cu035     18 24.09999 24.09999        NA    1
#> Evit200.551 6288 Cu035     18 24.09999 28.09999  6.200001    2
#> Evit200.552 6288 Cu035     18 24.09999 32.50000 15.900002    3
#> Evit200.553 6288 Cu035     18 24.09999 40.59998 30.800003    4
#> Evit200.554 6288 Cu035     18 24.09999 46.00000 40.900002    5
#> Evit200.555 6288 Cu035     18 24.09999 53.79999 52.100002    6
#> Evit200.598 6909 Cu035     13 24.70000 24.70000        NA    1
#> Evit200.599 6909 Cu035     13 24.70000 28.89999  5.900002    2
#> Evit200.600 6909 Cu035     13 24.70000 34.59998 15.000000    3
#> Evit200.601 6909 Cu035     13 24.70000 42.19998 26.600000    4
#> Evit200.602 6909 Cu035     13 24.70000 48.19998 38.000000    5
#> Evit200.603 6909 Cu035     13 24.70000 54.39999 50.900002    6
#> Evit200.622 6912 Cu000     13 24.50000 24.50000        NA    1
#> Evit200.623 6912 Cu000     13 24.50000 28.39999  7.100000    2
#> Evit200.624 6912 Cu000     13 24.50000 35.00000 17.300001    3
#> Evit200.625 6912 Cu000     13 24.50000 41.89999 29.500000    4
#> Evit200.626 6912 Cu000     13 24.50000 49.00000 40.900002    5
#> Evit200.627 6912 Cu000     13 24.50000 55.59998 54.200001    6
#> Evit200.670 8053 Cu175     20 21.89999 21.89999        NA    1
#> Evit200.671 8053 Cu175     20 21.89999 26.79999  9.600000    2
#> Evit200.672 8053 Cu175     20 21.89999 34.69998 21.800003    3
#> Evit200.673 8053 Cu175     20 21.89999 41.29999 36.700001    4
#> Evit200.674 8053 Cu175     20 21.89999 48.69998 51.800003    5
#> Evit200.675 8053 Cu175     20 21.89999 57.89999 71.400002    6
#> Evit200.694 8141 Cu035     23 24.20000 24.20000        NA    1
#> Evit200.695 8141 Cu035     23 24.20000 25.39999  3.900002    2
#> Evit200.696 8141 Cu035     23 24.20000 30.70000 13.300003    3
#> Evit200.697 8141 Cu035     23 24.20000 35.59998 24.600002    4
#> Evit200.698 8141 Cu035     23 24.20000 42.79999 38.500000    5
#> Evit200.699 8141 Cu035     23 24.20000 49.50000 53.000000    6
#> Evit200.700 8141 Cu035     23 24.20000 57.19998 67.600000    7
#> Evit200.754 8193 Cu035     21 29.29999 29.29999        NA    1
#> Evit200.755 8193 Cu035     21 29.29999 34.00000 10.500000    2
#> Evit200.756 8193 Cu035     21 29.29999 39.00000 20.200005    3
#> Evit200.757 8193 Cu035     21 29.29999 45.39999 34.800003    4
#> Evit200.758 8193 Cu035     21 29.29999 51.79999 51.000000    5
#> Evit200.759 8193 Cu035     21 29.29999 58.79999 70.000000    6
#> Evit200.766 8195 Cu000     21 25.29999 25.29999        NA    1
#> Evit200.767 8195 Cu000     21 25.29999 31.59999  9.600006    2
#> Evit200.768 8195 Cu000     21 25.29999 34.00000 19.600006    3
#> Evit200.769 8195 Cu000     21 25.29999 35.19998 28.000000    4
#> Evit200.770 8195 Cu000     21 25.29999 37.29999 41.000000    5
#> Evit200.771 8195 Cu000     21 25.29999 40.79999 57.400002    6
#> Evit200.772 8195 Cu000     21 25.29999 47.69998 69.000000    7
#> Evit200.773 8195 Cu000     21 25.29999 57.00000 86.300001    8
#> Evit200.778 8269 Cu175     22 22.09999 22.09999        NA    1
#> Evit200.779 8269 Cu175     22 22.09999 26.29999  6.000000    2
#> Evit200.780 8269 Cu175     22 22.09999 33.29999 16.800001    3
#> Evit200.781 8269 Cu175     22 22.09999 38.00000 27.400002    4
#> Evit200.782 8269 Cu175     22 22.09999 39.59998 36.200001    5
#> Evit200.783 8269 Cu175     22 22.09999 41.39999 43.300003    6
#> Evit200.784 8269 Cu175     22 22.09999 46.29999 54.000000    7
#> Evit200.785 8269 Cu175     22 22.09999 52.39999 70.000000    8
#> Evit200.838 8439 Cu175     24 27.29999 27.29999        NA    1
#> Evit200.839 8439 Cu175     24 27.29999 31.20000  8.800000    2
#> Evit200.840 8439 Cu175     24 27.29999 37.00000 20.000000    3
#> Evit200.841 8439 Cu175     24 27.29999 43.89999 33.200001    4
#> Evit200.842 8439 Cu175     24 27.29999 49.79999 44.800003    5
#> Evit200.843 8439 Cu175     24 27.29999 57.19998 58.000000    6